Make keepassxc-cli show return all matching entries.
As mentioned in issue #1967 show should return all matching items
instead of the first one. If a UUID is passed to show it will return one
entry if there is a match. For a path or name it will return all
matching entries.

TODO: Replace us of QMap.unite() to a function that skips duplicates.